+++ internal/cli/email_templates.go
@@ -24,12 +24,13 @@ const (
 	emailTemplateMFAEnrollment      = "mfa-enrollment"
 	emailTemplateMFACode            = "mfa-code"
 	emailTemplateUserInvitation     = "user-invitation"
+	emailTemplateAsyncApproval      = "async-approval"
 )
 
 var (
 	emailTemplateTemplate = Argument{
 		Name: "Template",
-		Help: fmt.Sprintf("Template name. Can be '%s', '%s', '%s', '%s', '%s', '%s', '%s', '%s', '%s' or '%s'",
+		Help: fmt.Sprintf("Template name. Can be '%s', '%s', '%s', '%s', '%s', '%s', '%s', '%s', '%s', '%s' or '%s'",
 			emailTemplateVerifyLink,
 			emailTemplateVerifyCode,
 			emailTemplateChangePassword,
@@ -39,7 +40,8 @@ var (
 			emailTemplatePasswordBreach,
 			emailTemplateMFAEnrollment,
 			emailTemplateMFACode,
-			emailTemplateUserInvitation),
+			emailTemplateUserInvitation,
+			emailTemplateAsyncApproval),
 	}
 
 	emailTemplateBody = Flag{
@@ -99,6 +101,7 @@ var (
 		{"Enroll in Multifactor Authentication", emailTemplateMFAEnrollment},
 		{"Verification Code for Email MFA", emailTemplateMFACode},
 		{"User Invitation", emailTemplateUserInvitation},
+		{"Async Approval", emailTemplateAsyncApproval},
 	}
 )
 
@@ -152,6 +155,7 @@ func showEmailTemplateCmd(cli *cli) *cobra.Command {
 	}
 
 	cmd.Flags().BoolVar(&cli.json, "json", false, "Output in json format.")
+	cmd.Flags().BoolVar(&cli.jsonCompact, "json-compact", false, "Output in compact json format.")
 
 	return cmd
 }
@@ -178,13 +182,15 @@ func updateEmailTemplateCmd(cli *cli) *cobra.Command {
 		Example: `  auth0 email templates update
   auth0 email templates update <template>
   auth0 email templates update <template> --json
+  auth0 email templates update <template> --json-compact
   auth0 email templates update welcome --enabled=true
   auth0 email templates update welcome --enabled=true --body "$(cat path/to/body.html)"
   auth0 email templates update welcome --enabled=false --body "$(cat path/to/body.html)" --from "welcome@example.com"
   auth0 email templates update welcome --enabled=true --body "$(cat path/to/body.html)" --from "welcome@example.com" --lifetime 6100
   auth0 email templates update welcome --enabled=false --body "$(cat path/to/body.html)" --from "welcome@example.com" --lifetime 6100 --subject "Welcome"
   auth0 email templates update welcome --enabled=true --body "$(cat path/to/body.html)" --from "welcome@example.com" --lifetime 6100 --subject "Welcome" --url "https://example.com"
-  auth0 email templates update welcome -e=true -b "$(cat path/to/body.html)" -f "welcome@example.com" -l 6100 -s "Welcome" -u "https://example.com" --json`,
+  auth0 email templates update welcome -e=true -b "$(cat path/to/body.html)" -f "welcome@example.com" -l 6100 -s "Welcome" -u "https://example.com" --json
+  auth0 email templates update welcome -e=true -b "$(cat path/to/body.html)" -f "welcome@example.com" -l 6100 -s "Welcome" -u "https://example.com" --json-compact`,
 		RunE: func(cmd *cobra.Command, args []string) error {
 			if len(args) > 0 {
 				inputs.Template = args[0]
@@ -284,6 +290,7 @@ func updateEmailTemplateCmd(cli *cli) *cobra.Command {
 	}
 
 	cmd.Flags().BoolVar(&cli.json, "json", false, "Output in json format.")
+	cmd.Flags().BoolVar(&cli.jsonCompact, "json-compact", false, "Output in compact json format.")
 	cmd.Flags().BoolVar(&cli.force, "force", false, "Skip confirmation.")
 	emailTemplateBody.RegisterStringU(cmd, &inputs.Body, "")
 	emailTemplateFrom.RegisterStringU(cmd, &inputs.From, "")
@@ -321,6 +328,8 @@ func apiEmailTemplateFor(v string) string {
 		return "mfa_oob_code"
 	case emailTemplateUserInvitation:
 		return "user_invitation"
+	case emailTemplateAsyncApproval:
+		return "async_approval"
 	default:
 		return v
 	}