+++ internal/cli/log_streams.go
@@ -35,6 +35,26 @@ var (
 		Help:         "The name of the log stream.",
 		AlwaysPrompt: true,
 	}
+
+	logStreamPIIConfig = Flag{
+		Name:      "PII Configuration",
+		LongForm:  "pii-config",
+		ShortForm: "c",
+		Help: "Specifies how PII fields are logged, Formatted as JSON. \n" +
+			"including which fields to log (first_name, last_name, username, email, phone, address)," +
+			"the protection method (mask or hash), and the hashing algorithm (xxhash). \n" +
+			" Example : " + `{"log_fields": ["first_name", "last_name"], "method": "mask", "algorithm": "xxhash"}. ` + "\n",
+		AlwaysPrompt: true,
+	}
+
+	logStreamFilters = Flag{
+		Name:      "Matching Filters",
+		LongForm:  "filters",
+		ShortForm: "m",
+		Help: "Events matching these filters will be delivered by the stream, Formatted as JSON. \n" +
+			"Example: " + `"[{"type":"category","name":"auth.login.fail"},{"type":"category","name":"auth.signup.fail"}]"`,
+		AlwaysPrompt: true,
+	}
 )
 
 func logStreamsCmd(cli *cli) *cobra.Command {
@@ -67,6 +87,7 @@ func listLogStreamsCmd(cli *cli) *cobra.Command {
 		Example: `  auth0 logs streams list
   auth0 logs streams ls
   auth0 logs streams ls --json
+  auth0 logs streams ls --json-compact
   auth0 logs streams ls --csv`,
 		RunE: func(cmd *cobra.Command, args []string) error {
 			var list []*management.LogStream
@@ -79,14 +100,14 @@ func listLogStreamsCmd(cli *cli) *cobra.Command {
 				return fmt.Errorf("failed to list log streams: %w", err)
 			}
 
-			cli.renderer.LogStreamList(list)
-			return nil
+			return cli.renderer.LogStreamList(list)
 		},
 	}
 
 	cmd.Flags().BoolVar(&cli.json, "json", false, "Output in json format.")
+	cmd.Flags().BoolVar(&cli.jsonCompact, "json-compact", false, "Output in compact json format.")
 	cmd.Flags().BoolVar(&cli.csv, "csv", false, "Output in csv format.")
-	cmd.MarkFlagsMutuallyExclusive("json", "csv")
+	cmd.MarkFlagsMutuallyExclusive("json", "json-compact", "csv")
 
 	return cmd
 }
@@ -104,7 +125,8 @@ func showLogStreamCmd(cli *cli) *cobra.Command {
 		Long:  "Display information about a log stream.",
 		Example: `  auth0 logs streams show
   auth0 logs streams show <log-stream-id>
-  auth0 logs streams show <log-stream-id> --json`,
+  auth0 logs streams show <log-stream-id> --json
+  auth0 logs streams show <log-stream-id> --json-compact`,
 		RunE: func(cmd *cobra.Command, args []string) error {
 			if len(args) == 0 {
 				err := logStreamID.Pick(cmd, &inputs.ID, cli.allLogStreamsPickerOptions)
@@ -124,12 +146,12 @@ func showLogStreamCmd(cli *cli) *cobra.Command {
 			}); err != nil {
 				return fmt.Errorf("failed to read log stream: %w", err)
 			}
-			cli.renderer.LogStreamShow(a)
-			return nil
+			return cli.renderer.LogStreamShow(a)
 		},
 	}
 
 	cmd.Flags().BoolVar(&cli.json, "json", false, "Output in json format.")
+	cmd.Flags().BoolVar(&cli.jsonCompact, "json-compact", false, "Output in compact json format.")
 
 	return cmd
 }